Moved Post.tsx to a different folder
