+++ apps/admin-x-activitypub/src/routes.tsx
@@ -8,7 +8,7 @@ import Onboarding from '@components/layout/Onboarding';
 import OnboardingStep1 from '@components/layout/Onboarding/Step1';
 import OnboardingStep2 from '@components/layout/Onboarding/Step2';
 import OnboardingStep3 from '@components/layout/Onboarding/Step3';
-import Post from './views/Feed/components/Post';
+import Post from './views/Feed/Post';
 import Profile from '@views/Profile';
 import ProfileRR from '@views/Profile/Profile-RR';
 import {Navigate, RouteObject} from '@tryghost/admin-x-framework';
